experiments setup for cloud
this changes clang compilation flags to target t4 and include C
libraries to recognize functions such as sqrt

+++ src/include/body.h
@@ -5,7 +5,7 @@
 
 static inline float my_rsqrtf(float x)
 {
-    return 1 / sqrt(x);
+    return 1.0f / (float)sqrt((double)x);
 }
 
 typedef struct
